+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Added new grid-width classes to grid component
   - Added better vertical-align for embedded content to base component
   - Added :before and :after to print component
+  - Added gridparallax component
   - Fixed Animated slideset filter
   - Fixed noDragClass option (sortable)
   - Fixed drop target detection (sortable)